fix warnings

+++ Lab3/TvSet/TvSet.cpp
@@ -107,7 +107,7 @@ bool CTvSet::SetChannelName(int number, string name)
 	return false;
 }
 
-int CTvSet::GetCurrChannel() const
+size_t CTvSet::GetCurrChannel() const
 {
 	return m_isOn ? m_currChannel : 0;
 }
@@ -123,7 +123,7 @@ bool CTvSet::DeleteChannelName(string name)
 	map<string, size_t>::iterator itName = channelName.find(name);
 	if (m_isOn && itName != channelName.end())
 	{
-		int number = itName->second;
+		size_t number = itName->second;
 		channelName.erase(itName);
 		channelNumber.erase(number);
 		return true;
